requirements.txt
Flask
PyPDF2
Pillow
Werkzeug
reportlab
